separate panel from panel bcast tests
